refactor(benchmark): Split the simulation module into milestones, presets, purchases, quests and snapshots, and derive milestone checks from one list.
